+++ setup.cfg
@@ -22,10 +22,10 @@ exclude = cmip6_preprocessing/_version.py
 
 
 [metadata]
-name = {{cookiecutter.project_name.lower().replace(' ', '_').replace('-', '_')}}
-description = 'Analysis ready CMIP6 data the easy way'
-author = 'cmip6_preprocessing developers'
-url = https://github.com/{{cookiecutter.github_username}}/{{cookiecutter.project_name.lower().replace(' ', '_')}}
+name = "cmip6_preprocessing"
+description = "Analysis ready CMIP6 data the easy way"
+author = "cmip6_preprocessing developers"
+url="https://github.com/jbusecke/cmip6_preprocessing"
 long_description = file: README.md
 long_description_content_type = text/markdown
 license = "MIT"
@@ -35,7 +35,7 @@ license_file = LICENSE.txt
 # For details see: https://pypi.org/classifiers/
 
 classifiers =
-    Development Status :: 5 - Production/Stable 
+    Development Status :: 4 - Beta 
     Topic :: Scientific/Engineering
     Intended Audience :: Science/Research
     Operating System :: OS Independent
@@ -45,17 +45,23 @@ classifiers =
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     # Dont change this one
-    License :: OSI Approved ::{% if cookiecutter.open_source_license == 'MIT' %} MIT License {% elif cookiecutter.open_source_license == 'BSD-3-Clause' %} BSD License {% endif %}
+    License :: OSI Approved :: MIT
 
 ## Add your email here
-author_email = AUTHOR@EMAIL.COM
+author_email = jbusecke@princeton.edu
 
 
 ### make sure to fill in your dependencies!
 [options]
 install_requires =
     numpy
-    requests
+    xarray
+    pandas
+    pyproj
+    matplotlib
+    regionmask
+    xgcm
+    cftime
 setup_requires= 
     setuptools_scm
 python_requires = >=3.6
@@ -63,22 +69,3 @@ python_requires = >=3.6
 
 zip_safe = False
 packages = find:
-
-
-# setup(
-#     name="cmip6_preprocessing",
-#     version=versioneer.get_version(),
-#     cmdclass=versioneer.get_cmdclass(),
-#     license="MIT",
-#     author_email="jbusecke@princeton.edu",
-#     url="https://github.com/jbusecke/cmip6_preprocessing",
-#     packages=["cmip6_preprocessing"],
-#     include_package_data=True,
-#     install_requires=requirements,
-#     keywords="cmip6_preprocessing",
-#     classifiers=[
-#         "Programming Language :: Python :: 2.7",
-#         "Programming Language :: Python :: 3.6",
-#         "Programming Language :: Python :: 3.7",
-#     ],
-# )
